ItemData.java to change age to absolute year to calculate the year the item was bought. ViewItemActivity.java to update the itemData so that the item is updated when user comes back to the Edit activity from the View activity. EditItemActivity.java to fix the error display the Age, index must be typecasted to int.

+++ src/com/example/closetstylist/EditItemActivity.java
@@ -330,7 +330,7 @@ public class EditItemActivity extends Activity {
 		tempMin.setSelection(temperatureArray.indexOf(Integer.toString(item.getTempMin())));
 		category.setSelection(categoryArray.indexOf(item.getCategory()));
 		brand.setSelection(brandArray.indexOf(item.getBrand()));
-		age.setSelection(ageArray.indexOf(Double.toString(item.getAge())));
+		age.setSelection(ageArray.indexOf(Integer.toString((int)item.getAge())));
 		material.setSelection(materialArray.indexOf(item.getMaterial()));
 	}
